Accept file-format version 3, as well as 1
This appears to mean nearly the same as version 1.  In particular I
believe the meaning of this is such that an old reader which doesn't
know about it can treat the file as if it were version 1, and get all
the same information from it that it could if it had actually been
written according to version 1.

Here's the current version-number definitions:

  // https://chromium.googlesource.com/chromium/src.git/+/refs/tags/89.0.4389.90/components/sessions/core/command_storage_backend.cc

  // File version number.
  constexpr int32_t kFileVersion1 = 1;
  constexpr int32_t kEncryptedFileVersion = 2;
  // The versions that are used if `use_marker` is true.
  constexpr int32_t kFileVersionWithMarker = 3;
  constexpr int32_t kEncryptedFileVersionWithMarker = 4;

The "marker" bit is important for *new* readers, to tell them to look
out for a particular command type in the sequence, and if it's missing
to infer the file is incomplete.  But old readers that weren't looking
for that marker anyway can just proceed as usual.